feat(CartDrawer): extract cart drawer logic into a custom hook for improved separation of concerns
